fix: rename cms folders
