+++ src/components/app/builder.tsx
@@ -5,7 +5,7 @@ import { BuilderComponent, useIsPreviewing } from '@builder.io/react'
 import { notFound } from 'next/navigation'
 
 import { useHasHydrated } from '@/hooks/useHasHydrated'
-import { maybeInitCMSClient } from '@/utils/web/clientCMS'
+import { maybeInitCMSClient } from '@/utils/web/builder/clientCMS'
 
 type BuilderPageProps = ComponentProps<typeof BuilderComponent>
