+++ src/utils/server/builder/models/data/questionnaire.ts
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/nextjs'
 import pRetry from 'p-retry'
 
-import { serverCMS } from '@/utils/server/serverCMS/serverCMS'
+import { serverCMS } from '@/utils/server/builder/serverCMS'
 import { zodQuestionnaireSchemaValidation } from '@/utils/shared/getSWCQuestionnaire'
 import { getLogger } from '@/utils/shared/logger'
